breaking down the components of the project
