+++ setup.py
@@ -1,8 +1,8 @@
-from setuptools import setup, find_packages
+from setuptools import setup,find_packages
 
 setup(
   name='codemon',
-  version='0.0.1',
+  version='0.0.2',
   author='Ankush Bhardwaj',
   packages=find_packages(),
   install_requires=[
